Various Minor Changes to make gcc HAppy.
git-svn-id: svn://svn.code.sf.net/p/eapidk/code/trunk@72 3723b92f-f4ce-431d-9be9-3c706c148987

+++ apps/EeePProg/Str2Value.c
@@ -897,7 +897,7 @@ AssignValue_VAB(
     unsigned long ulMask=(((unsigned long )1<<siBitLen)-1);
     EAPI_APP_RETURN_ERROR_IF_S(
         AssignValue_VAB, 
-        sizeof(unsigned long)*8<siBitOffset+siBitLen,
+        sizeof(unsigned long)*8<(unsigned)(siBitOffset+siBitLen),
         EAPI_STATUS_INVALID_PARAMETER
       );
     EAPI_APP_RETURN_ERROR_IF_S(
@@ -993,7 +993,7 @@ RecoverValue_VAB(
     unsigned long ulMask=(((unsigned long)1<<siBitLen)-1);
     EAPI_APP_RETURN_ERROR_IF_S(
         RecoverValue_VAB, 
-        sizeof(unsigned long)*8<siBitOffset+siBitLen,
+        sizeof(unsigned long)*8<(unsigned)(siBitOffset+siBitLen),
         EAPI_STATUS_INVALID_PARAMETER
       );
     *(unsigned long long*)psllValue=(*pulValue>>siBitOffset)&ulMask;
